+++ src/Podenv/Main.hs
@@ -65,7 +65,7 @@ main = do
             else do
                 let doUpdate = Podenv.Runtime.updateRuntime run (ar ^. arApplication . appRuntime)
                 when update doUpdate
-                Podenv.Runtime.execute run Foreground ctx
+                void $ Podenv.Runtime.execute run Foreground ctx
                 maybeUpdate ar run gl doUpdate
 
 maybeUpdate :: ApplicationResource -> RunEnv -> GlobalEnv -> ReaderT GlobalEnv IO () -> ReaderT GlobalEnv IO ()